+++ NEWS.md
@@ -1,5 +1,12 @@
 # ChromatogramsVis 0.1
 
+## Changes in 0.1.8
+
+- Update visibility sidebar menu
+- Export filtered MsExperiment
+- Fix bug export filtered object
+- Increase limit input file from 5Mb to 300Mb
+
 ## Changes in 0.1.7
 
 - Complete support of MsExperiment in Galaxy